requirements.txt
uvicorn~=0.38.0
fastapi~=0.120.2
SQLAlchemy~=2.0.44
pydantic~=2.12.3
httpx
python-dotenv